requirements.txt
Django
psycopg2

django-ckeditor
Pillow
django-filer

praw
